longinregister can link server
